moving jokes function to separate component
